fixed testimony
